requirements.txt
google-generativeai
pinecone-client
langchain
pypdf
python-dotenv